move query from StoriesList to index component
